requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
alembic==1.17.2
pydantic-settings==2.12.0
psycopg2-binary==2.9.5
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic[email]
python-multipart==0.0.6